fixing stuff

+++ models/level.go
@@ -1,6 +1,6 @@
 package models
 
 type Level struct {
-	Value int32   `json:"value" firestore:"value"`
-	Exp   float64 `json:"exp" firestore:"exp"`
+	Value int32 `json:"value" firestore:"value"`
+	Exp   int32 `json:"exp" firestore:"exp"`
 }
